% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{adj_spec_test}
\alias{adj_spec_test}
\title{Adjusted spectral test}
\usage{
adj_spec_test(A, K, z = NULL, DC = T, theta = NULL, B = NULL, ...)
}
\arguments{
\item{A}{adjacency matrix}

\item{K}{number of communities}

\item{z}{label vector for rows of adjacency matrix. If not given, will be calculated by
the spectral clustering}

\item{DC}{whether or not include degree correction in the parameter estimation}

\item{theta}{give the propensity parameter directly}

\item{B}{give the connectivity matrix directly}

\item{...}{parameters for spectral clustering.}
}
\value{
A list of result
\item{stat}{NAC or NAC+ test statistic.}
\item{z}{row label vector.}
\item{y}{column label vector.}
}
\description{
Adjusted spectral test
}
