% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_select.R
\name{bethe_hessian_select}
\alias{bethe_hessian_select}
\title{Estimate the community number using the Beth Hessian matrix}
\usage{
bethe_hessian_select(A, Kmax)
}
\arguments{
\item{A}{adjacency matrix.}

\item{Kmax}{the maximum number of communities to check.}
}
\value{
A list of result
\item{K}{estimated community number}
\item{rho}{eigenvalues of the Beth-Hessian matrix}
}
\description{
Estimates the number of communities under block models by using the spectral
properties of network Beth-Hessian matrix with moment correction.
}
