% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{compute_block_sums}
\alias{compute_block_sums}
\title{Block sum of an adjacency matrix}
\usage{
compute_block_sums(A, z)
}
\arguments{
\item{A}{adjacency matrix.}

\item{z}{label vector.}
}
\value{
A K x L matrix with (k,l)-th element as \eqn{sum_{i,j} A_{i,j} 1{z_i = k, z_j = l}}
}
\description{
Compute the block sum of an adjacency matrix given a label vector.
}
\keyword{estimation}
