% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_confusion_matrix}
\alias{compute_confusion_matrix}
\title{Compute confusion matrix}
\usage{
compute_confusion_matrix(z, y, K = NULL)
}
\arguments{
\item{z}{a label vector}

\item{y}{a label vector}

\item{K}{number of labels in both \code{z} and \code{y}}
}
\value{
A \code{K}x\code{K} confusion matrix between \code{z} and \code{y}
}
\description{
Compute confusion matrix
}
\keyword{evaluation}
