% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_mutual_info}
\alias{compute_mutual_info}
\title{Compute normalized mutual information (NMI)}
\usage{
compute_mutual_info(z, y)
}
\arguments{
\item{z}{a label vector}

\item{y}{a label vector}
}
\description{
Compute normalized mutual information (NMI)
}
\keyword{evaluation}
