% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{computeBlockSums}
\alias{computeBlockSums}
\title{Block sum of an adjacency matrix}
\usage{
computeBlockSums(A, z)
}
\arguments{
\item{A}{adjacency matrix.}

\item{z}{label vector.}
}
\value{
matrix $B\link{k,l} = sum_{i,j} A\link{i,j} 1{z_i = k, z_j = l}$
}
\description{
Compute the block sum of an adjacency matrix given a label vector.
}
\keyword{estimation}
