% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estim_dcsbm}
\alias{estim_dcsbm}
\title{Estimate model parameters of a DCSBM}
\usage{
estim_dcsbm(A, z)
}
\arguments{
\item{A}{adjacency matrix.}

\item{z}{label vector.}
}
\value{
A list of result
\item{B}{estimated connectivity matrix.}
\item{theta}{estimated node propensity parameter.}
}
\description{
Compute the block sum of an adjacency matrix given a label vector.
}
\details{
\deqn{\hat B_{k\ell} = \frac{N_{k\ell}(\hat z)}{m_{k\ell} (\hat z)}, \quad \hat \theta_i =  \frac{n_{\hat z_i}(\zh) d_i}{\sum_{j : \hat z_j = \hat z_i} d_i}, \quad \hat \pi_k = n_k(\zh) / n}
}
\keyword{estimation}
