% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{eval_dcsbm_bic}
\alias{eval_dcsbm_bic}
\title{compute BIC score}
\usage{
eval_dcsbm_bic(A, z, K, poi)
}
\arguments{
\item{A}{adjacency matrix}

\item{z}{label vector}

\item{K}{number of community in \code{z}}

\item{poi}{whether to use Poisson version of likelihood}
}
\description{
compute BIC score when fitting a DCSBM to network data
}
\details{
the BIC score is calculated by log likelihood minus \eqn{K*(K + 1)*log(n)/2}
}
\keyword{mod_sel}
