% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{eval_dcsbm_loglr}
\alias{eval_dcsbm_loglr}
\title{Log likelihood of a DCSBM}
\usage{
eval_dcsbm_loglr(A, labels, poi = T, eps = 1e-06)
}
\arguments{
\item{A}{adjacency matrix}

\item{labels}{a matrix with two columns as two different label vectors}

\item{poi}{whether to use Poisson version of likelihood}

\item{eps}{truncation threshold for the Bernoulli likelihood,
used when parameter phat is close to 1 or 0.}
}
\value{
log likelihood ratio
}
\description{
Compute the log likelihood of a DCSBM, using estimated parameters
B, theta based on the given label vector
}
\details{
log likelihood ratio between fitting DCSBM with the label vectors
as the second and first column of \code{labels}
}
\keyword{mod_sel}
