% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_utils.R
\name{extract_largest_cc}
\alias{extract_largest_cc}
\title{Extract largest component}
\usage{
extract_largest_cc(gr, mode = "weak")
}
\arguments{
\item{gr}{The network as an igraph object}

\item{mode}{Type of connected component ("weak"|"strong")}
}
\value{
An igraph object
}
\description{
Extract the largest connected component of a network
}
\keyword{utils}
