% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_utils.R
\name{extract_low_deg_comp}
\alias{extract_low_deg_comp}
\title{Extract low-degree component}
\usage{
extract_low_deg_comp(g, deg_prec = 0.75, verb = FALSE)
}
\arguments{
\item{g}{The network as an igraph object}

\item{deg_prec}{The cut-off degree percentile}

\item{verb}{Whether to be verbose (TRUE|FALSE)}
}
\value{
An igraph object
}
\description{
Extract a low-degree connected component of a network
}
\keyword{utils}
