% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpl.R
\name{fastCPL}
\alias{fastCPL}
\title{CPL}
\usage{
fastCPL(Amat, Kcommunities, ilabels = NULL, Iter = 10)
}
\arguments{
\item{Amat}{adjacency matric}

\item{Kcommunities}{desired community number}

\item{ilabels}{initial label vector}

\item{Iter}{number of iterations}
}
\value{
estimated community label vector of \code{Amat}
}
\description{
CPL
}
\keyword{comm_detect}
