% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fastSBM}
\alias{fastSBM}
\title{Sample SBM}
\usage{
fastSBM(z, Pmat)
}
\arguments{
\item{z}{Node labels (n x 1)}

\item{Pmat}{Connectivity matrix (K x K)}
}
\description{
Sample an adjacency matrix from a stochastic block model (SBM)
}
