% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_dcsbm_exav_deg}
\alias{get_dcsbm_exav_deg}
\title{Calculate the expected average degree of a DCSBM}
\usage{
get_dcsbm_exav_deg(n, pri, B, ex_theta = 1)
}
\arguments{
\item{n}{number of nodes}

\item{pri}{distribution of node labels (K x 1)}

\item{B}{connectivity matrix (K x K)}

\item{ex_theta}{expected value of theta}
}
\value{
expected average degree of a DCSBM
}
\description{
Calculate the expected average degree of a DCSBM
}
\keyword{models}
