% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_mat2vec}
\alias{label_mat2vec}
\title{Convert label matrix to vector}
\usage{
label_mat2vec(Z)
}
\arguments{
\item{Z}{a cluster assignment matrix}
}
\value{
A label vector that follows the assignment matrix
}
\description{
Convert label matrix to vector
}
\keyword{utils}
