% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nac_test.R
\name{nac_test}
\alias{nac_test}
\title{NAC test}
\usage{
nac_test(A, K, z = NULL, y = NULL, plus = T, ...)
}
\arguments{
\item{A}{adjacency matrix}

\item{K}{number of communities}

\item{z}{label vector for rows of adjacency matrix. If not given, will be calculated by
the spectral clustering}

\item{y}{label vector for columns of adjacency matrix.}

\item{plus}{whether or not use column label vector with (K+1) communities, default is TRUE}

\item{...}{parameters for spectral clustering.}
}
\value{
A list of result
\item{stat}{NAC or NAC+ test statistic.}
\item{z}{row label vector.}
\item{y}{column label vector.}
}
\description{
NAC and NAC+ test
}
