% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_deg_dist}
\alias{plot_deg_dist}
\title{Plot digree distribution}
\usage{
plot_deg_dist(gr, logx = T)
}
\arguments{
\item{gr}{the network as an igraph object}
}
\description{
Plot the digree distribution of a network on log scale
}
