% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_net}
\alias{plot_net}
\title{Plot a network}
\usage{
plot_net(
  gr,
  community = NULL,
  color_map = NULL,
  extract_lcc = TRUE,
  heavy_edge_deg_perc = 0.97,
  coord = NULL,
  vsize_func = function(deg) log(deg + 3) * 1,
  vertex_border = F,
  niter = 1000,
  vertex_alpha = 0.4,
  remove_loops = T,
  make_simple = F,
  show_legend = F,
  ...
)
}
\arguments{
\item{gr}{the network as an igraph object}

\item{community}{community assignment; vector of node labels}

\item{...}{other settings}
}
\description{
Plot a network using degree-modulated node sizes, community colors and other enhancements
}
