% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curves}
\usage{
plot_roc(roc_results, method_names = NULL)
}
\arguments{
\item{roc_results}{data frame \code{roc} from the output list of \code{\link{simulate_roc}}}
}
\description{
Plot ROC curves given results from \code{\link{simulate_roc}}
}
\keyword{plotting}
