% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_profiles.R
\name{plot_smooth_profile}
\alias{plot_smooth_profile}
\title{Plot community profiles}
\usage{
plot_smooth_profile(
  tstat,
  net_name = "",
  trunc_type = "none",
  spar = 0.3,
  plot_null_spar = T,
  alpha = 0.3,
  base_font_size = 12
)
}
\description{
Plot the smooth community profiles based on a resampled statistic
}
