% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polblogs}
\alias{polblogs}
\title{Political blogs network}
\format{
An igraph data with 1490 nodes and 19090 edges
}
\usage{
data(polblogs)
}
\description{
This is a directed network of hyperlinks between political blogs about politics
in the United States of America.
}
\references{
\href{http://konect.cc/networks/dimacs10-polblogs/}{Data source}.
Original paper:\href{https://dl.acm.org/doi/10.1145/1134271.1134277}{The political blogosphere and the 2004 US election: divided they blog},
Adamic, Lada A., and Natalie Glance. "." Proceedings of the 3rd international workshop on Link discovery. 2005.
}
\keyword{datasets}
