% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printf}
\alias{printf}
\title{The usual "printf" function}
\usage{
printf(...)
}
\arguments{
\item{...}{printing object}
}
\value{
the value of the printing object
}
\description{
The usual "printf" function
}
\keyword{utils}
