% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsymperm}
\alias{rsymperm}
\title{Generate random symmetric permutation matrix}
\usage{
rsymperm(K)
}
\arguments{
\item{K}{size of the matrix}
}
\value{
A random K x K symmetric permutation matrix
}
\description{
Generate a random symmetric permutation matrix (recursively)
}
\keyword{utils}
