% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sample_dcer}
\alias{sample_dcer}
\title{Sample from a DCER}
\usage{
sample_dcer(theta)
}
\arguments{
\item{theta}{Node connectivity propensity vector (\eqn{n * 1})}
}
\value{
An adjacency matrix following DCSBM
}
\description{
Sample an adjacency matrix from a degree-corrected Erdős–Rényi model (DCER).
}
