% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sample_dcpp}
\alias{sample_dcpp}
\title{Sample from a DCPP}
\usage{
sample_dcpp(
  n,
  lambda,
  K,
  oir,
  theta = NULL,
  pri = rep(1, K)/K,
  normalize_theta = F
)
}
\description{
Sample from a degree-corrected planted parition model
}
