% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sample_dcsbm}
\alias{sample_dcsbm}
\title{Sample DCSBM}
\usage{
sample_dcsbm(z, Pmat, theta = 1)
}
\arguments{
\item{z}{Node labels (n x 1)}

\item{Pmat}{Connectivity matrix (K x K)}

\item{theta}{Node connectivity propensity vector (n x 1)}
}
\value{
an adjacency matrix following DCSBM
}
\description{
Sample an adjacency matrix from a degree-corrected block model (DCSBM)
}
