% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nac_test.R
\name{snac_select}
\alias{snac_select}
\title{Estimate community number with SNAC+}
\usage{
snac_select(A, Kmin = 1, Kmax, alpha, labels = NULL)
}
\arguments{
\item{A}{adjacency matrix}

\item{Kmin}{minimum candidate community number}

\item{Kmax}{maximum candidate community number}

\item{alpha}{significance level for rejecting the null hypothsis}

\item{labels}{a matrix with each column being a row label vector for a
candidate community number}
}
\value{
estimated community number
}
\description{
Applying SNAC+ test sequentially to estimate community number of a network
fit to DCSBM
}
\seealso{
\link{snac_test}
}
\keyword{mod_sel}
