% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nac_test.R
\name{snac_test}
\alias{snac_test}
\title{SNAC test}
\usage{
snac_test(
  A,
  K,
  z = NULL,
  ratio = 0.5,
  fromEachCommunity = TRUE,
  plus = TRUE,
  niter = 1,
  ...
)
}
\arguments{
\item{A}{adjacency matrix}

\item{K}{desired number of communities}

\item{z}{label vector for rows of adjacency matrix. If not provided, will be estimated by
\link{spec_clust}}

\item{ratio}{ratio of subsampled nodes from the network}

\item{fromEachCommunity}{whether subsample from each estimated community or the full network,
default is TRUE}

\item{plus}{whether or not use column label vector with (K+1) communities, default is TRUE}

\item{niter}{number of times the statisitcs are computed}

\item{...}{parameters for spectral clustering.}
}
\value{
A list of result
\item{stat}{SNAC or SNAC+ test statistic.}
\item{z}{row label vector.}
}
\description{
The SNAC test to measure the goodness-of-fit of the DCSBM to network data.

The function computes the SNAC+ or SNAC statistics in the paper below.
The row label vector of the adjacency matrix could be given through \code{z} otherwise will
be estimated by \link{spec_clust}. One can specify the ratio of nodes used to estimate column
label vector. If \code{plus = TRUE}, the column labels will be estimated by \link{spec_clust} with
(\code{K}+1) clusters, i.e. performing SNAC+ test, otherwise with \code{K} clusters SNAC test.
One can also get multiple test statistics with repeated random subsampling on nodes.
}
\section{References}{
 \href{https://arxiv.org/abs/2012.15047}{Adjusted chi-square test for degree-corrected block models},
Linfan Zhang, Arash A. Amini, arXiv preprint arXiv:2012.15047, 2020.
}

\examples{
A <- sample_dcpp(500, 10, 4, 0.1)
snac_test(A, K = 4)$stat

}
