% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_select.R
\name{bethe_hessian_select}
\alias{bethe_hessian_select}
\title{Beth-Hessian model selection}
\usage{
bethe_hessian_select(A, Kmax)
}
\arguments{
\item{A}{adjacency matrix.}

\item{Kmax}{the maximum number of communities to check.}
}
\value{
A list of result
\item{K}{estimated the number of communities}
\item{rho}{eigenvalues of the Beth-Hessian matrix}
}
\description{
Estimate the number of communities under block models by using the spectral
properties of network Beth-Hessian matrix with moment correction.
}
\section{References}{

\href{https://arxiv.org/abs/1507.00827}{Estimating the number of communities in networks by spectral methods},
Can M. Le, Elizaveta Levina, arXiv preprint arXiv:1507.00827, 2015
}

\keyword{mod_sel}
