% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{eval_dcsbm_like}
\alias{eval_dcsbm_like}
\title{Log likelihood of a DCSBM}
\usage{
eval_dcsbm_like(A, z, poi = T, eps = 1e-06)
}
\arguments{
\item{A}{adjacency matrix}

\item{z}{label vector}

\item{poi}{whether to use Poisson version of likelihood}

\item{eps}{truncation threshold for the Bernoulli likelihood,
used when parameter phat is close to 1 or 0.}
}
\value{
log likelihood of a DCSBM
}
\description{
Compute the log likelihood of a DCSBM, using estimated parameters
B, theta based on the given label vector
}
\keyword{estimation}
