% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_vec2mat}
\alias{label_vec2mat}
\title{Convert label vector to matrix}
\usage{
label_vec2mat(z, K = NULL, sparse = FALSE)
}
\arguments{
\item{z}{a label vector}

\item{K}{number of labels in \code{z}}

\item{sparse}{whether the output should be sparse matrix}
}
\value{
A cluster assignment matrix that follows from the label vector \code{z}
}
\description{
Convert label vector to matrix
}
\keyword{utils}
