% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nac_test.R
\name{nac_test}
\alias{nac_test}
\title{NAC test}
\usage{
nac_test(A, K, z = NULL, y = NULL, plus = T, ...)
}
\arguments{
\item{A}{adjacency matrix}

\item{K}{number of communities}

\item{z}{label vector for rows of adjacency matrix. If not given, will be calculated by
the spectral clustering}

\item{y}{label vector for columns of adjacency matrix.}

\item{plus}{whether or not use column label vector with (\code{K}+1) communities, default is TRUE}

\item{...}{parameters for spectral clustering.}
}
\value{
A list of result
\item{stat}{NAC or NAC+ test statistic.}
\item{z}{row label vector.}
\item{y}{column label vector.}
}
\description{
The NAC test to measure the goodness-of-fit of the DCSBM to network data.

The function computes the NAC+ or NAC statistics in the paper below.
}
\section{References}{
 \href{https://arxiv.org/abs/2012.15047}{Adjusted chi-square test for degree-corrected block models},
Linfan Zhang, Arash A. Amini, arXiv preprint arXiv:2012.15047, 2020.
}

\examples{
A <- sample_dcpp(500, 10, 4, 0.1)
nac_test(A, K = 4)$stat
}
\seealso{
\link{snac_test}
}
