% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_deg_dist}
\alias{plot_deg_dist}
\title{Plot degree distribution}
\usage{
plot_deg_dist(gr, logx = TRUE)
}
\arguments{
\item{gr}{the network as an igraph object}

\item{logx}{whether the degree is in log scale.}
}
\value{
Histogram of the degree of 'gr'.
}
\description{
Plot the degree distribution of a network on log scale
}
