% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curves}
\usage{
plot_roc(roc_results, method_names = NULL, font_size = 16)
}
\arguments{
\item{roc_results}{data frame \code{roc} from the output list of \link{simulate_roc}}

\item{method_names}{a list of method names}

\item{font_size}{font size of the plot}
}
\value{
Roc plot based on results from \link{simulate_roc}
}
\description{
Plot ROC curves given results from \link{simulate_roc}.
}
\keyword{plotting}
