% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sample_dclvm}
\alias{sample_dclvm}
\title{Sample from a DCLVM}
\usage{
sample_dclvm(z, lambda, theta, npairs = NULL)
}
\description{
Sample form a degree-corrected latent variable model with Gaussian kernel
}
