% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nac_test.R
\name{snac_test}
\alias{snac_test}
\title{SNAC test}
\usage{
snac_test(
  A,
  K,
  z = NULL,
  ratio = 0.5,
  plus = TRUE,
  fromEachCommunity = TRUE,
  ...
)
}
\arguments{
\item{A}{adjacency matrix}

\item{K}{number of communities}

\item{z}{label vector for rows of adjacency matrix. If not given, will be calculated by}

\item{ratio}{ratio of subsampled nodes from the network}

\item{plus}{whether or not use column label vector with (K+1) communities, default is TRUE}

\item{fromEachCommunity}{whether subsample from each estimated community or the full network,
default is TRUE}

\item{...}{parameters for spectral clustering.}
}
\value{
SNAC or SNAC+ test statistics and label vector z and y.
}
\description{
perform SNAC test
}
